修改数据上报格式
